- Moves 'entertainment' story template to 'weekly-entertainment'.
