+++ packages/stickers/src/weekly-entertainment/entertainmentStar.js
