+++ packages/stickers/src/weekly-entertainment/index.js
