+++ packages/stickers/src/weekly-entertainment/musicNote.js
