+++ packages/templates/src/raw/weekly-entertainment/template.json
